# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Backend Pleno - Soluct API",
        "description": "Documentação da API RESTful para o desafio técnico da Soluct.\n * * **Observações Importantes:**\n * - Todas as requisições que enviam dados no corpo (POST, PUT, PATCH) devem usar o header `Content-Type: application/json`.\n * - Todas as requisições devem incluir o header `Accept: application/json` para garantir que a resposta seja recebida no formato JSON correto.\n * - Os endpoints que requerem autenticação estão marcados com um ícone de cadeado.",
        "contact": {
            "email": "elizaelnycolas@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza o login do usuário",
                "description": "Use um dos seguintes usuários de teste com a senha `senha123`:\n * \n * - joao.silva@example.com\n * - maria.oliveira@example.com\n * - carlos.santos@example.com\n * - ana.pereira@example.com\n * - lucas.almeida@example.com\n * ",
                "operationId": "f50ce14f5b1f1a59f3e85bdcf0b8918d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao.silva@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login realizado com sucesso."
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciais inválidas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao tentar realizar o login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocorreu um erro ao tentar realizar o login."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Mensagem de erro detalhada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza o logout do usuário autenticado",
                "operationId": "203b655d94610a1072245bd6098ec0ef",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Não foi possível realizar o logout.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não foi possível realizar o logout."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao tentar realizar o logout.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocorreu um erro ao tentar realizar o logout."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Mensagem de erro detalhada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}